Do refactoring
